Insert initial data by Squeryl

+++ app/models/Inflation.scala
@@ -3,4 +3,4 @@ package models
 import org.squeryl._
 import org.squeryl.PrimitiveTypeMode._
 
-case class Inflation (val id: Long = 0, val currency: String, val year: Int, val rate: Long) extends KeyedEntity[Long]
+case class Inflation (val id: Long = 0, val currency: String, val year: Int, val rate: Float) extends KeyedEntity[Long]
